hotfix stuff

+++ models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, String, BigInteger
 from sqlalchemy.ext.declarative import declarative_base
 Base = declarative_base()
 
@@ -7,6 +7,6 @@ class User(Base):
 
     __tablename__ = "user"
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     name = Column(String(200), nullable=False)
     password = Column(String(20), nullable=False)
